Adds AFNetworking as a dependency; Completes refactoring

+++ Wazza-iOS-SDK/WZSecurityService.m
@@ -83,7 +83,7 @@
 -(NSString*)hashContent:(NSString*)input {
     const char* str = [input UTF8String];
     unsigned char result[CC_SHA256_DIGEST_LENGTH];
-    CC_SHA256(str, strlen(str), result);
+    CC_SHA256(str, (unsigned int)strlen(str), result);
     
     NSMutableString *ret = [NSMutableString stringWithCapacity:CC_SHA256_DIGEST_LENGTH*2];
     for(int i = 0; i<CC_SHA256_DIGEST_LENGTH; i++)
